pyproject.toml
[project]
name = "musonius"
version = "0.1.0"
description = "Spec-driven development orchestrator — the AI coding multiplier"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Musonius Contributors" },
]
keywords = ["ai", "coding", "orchestrator", "cli", "developer-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "tree-sitter>=0.23",
    "tree-sitter-python>=0.23",
    "litellm>=1.50",
    "tiktoken>=0.7",
    "gitpython>=3.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "networkx>=3.0",
    "fastmcp>=0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.3",
    "mypy>=1.8",
    "types-PyYAML>=6.0",
    "types-networkx>=3.0",
]

[project.scripts]
musonius = "musonius.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "TC001",  # move application import into TYPE_CHECKING block
    "TC002",  # move third-party import into TYPE_CHECKING block
    "TC003",  # move stdlib import into TYPE_CHECKING block
]

[tool.ruff.lint.isort]
known-first-party = ["musonius"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["networkx.*", "yaml.*", "litellm.*", "tree_sitter.*", "tree_sitter_python.*", "fastmcp.*"]
ignore_missing_imports = true
